refactor: separate data query logic from server action
